server/requirements.txt
annotated-types==0.7.0
aspose-words==24.7.0
blinker==1.8.2
blobfile==2.1.1
cachetools==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
huggingface-hub==0.24.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
PyYAML==6.0.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rsa==4.9
tiktoken==0.1.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
XlsxWriter==3.2.0
